+++ CHANGES.rst
@@ -5,6 +5,7 @@ grafana-wtf changelog
 
 in progress
 ===========
+- Rename subcommand ``datasource-breakdown`` to ``explore datasources``
 
 2021-12-10 0.11.1
 =================
@@ -15,8 +16,8 @@ in progress
 - Upgrade to ``colored==1.4.3``. Thanks, @dslackw!
 - Tests: Use ``.env`` file for propagating environment variables to Docker Compose
 - CI/GHA test matrix: Use Grafana 7.5.11 and 8.3.1 and add Python 3.10
-- Add feature to display datasource breakdown, specifically for finding unused
-  data sources. Thanks, @chenlujjj!
+- Add feature to explore datasources, specifically for finding unused ones.
+  Thanks, @chenlujjj!
 
 2021-10-01 0.10.0
 =================